move postPath to supabase/post
