+++ web/components/posts/create-post.tsx
@@ -5,7 +5,7 @@ import { TextEditor, useTextEditor } from 'web/components/widgets/editor'
 import { createPost } from 'web/lib/firebase/api'
 import Router from 'next/router'
 import { MAX_POST_TITLE_LENGTH } from 'common/post'
-import { postPath } from 'web/lib/firebase/posts'
+import { postPath } from 'web/lib/supabase/post'
 import { Group } from 'common/group'
 import { ExpandingInput } from '../widgets/expanding-input'
 import { Button } from '../buttons/button'
