+++ web/pages/api/v0/reports.ts
@@ -11,7 +11,7 @@ import {
   listAllCommentsOnPost,
 } from 'web/lib/firebase/comments'
 import { ENV_CONFIG } from 'common/envs/constants'
-import { postPath } from 'web/lib/firebase/posts'
+import { postPath } from 'web/lib/supabase/post'
 import { getPost } from 'web/lib/supabase/post'
 import { richTextToString } from 'common/util/parse'
 import { getUser } from 'web/lib/firebase/users'
